requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
click==6.7
cycler==0.10.0
Flask==0.12.2
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.2
pillow==5.1.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
recordtype==1.1
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
urlparse2==1.1.1
Werkzeug==0.14.1
